add product edit function for admin

+++ client/E-commerce/src/pages/UserList.jsx
@@ -28,7 +28,9 @@ const UserList = () => {
   }, [dispatch, navigate, successDelete, userInfo]);
 
   const deleteUserHandle = (id) => {
-    dispatch(userDeleteAction(id));
+    if (window.confirm("Are you sure")) {
+      dispatch(userDeleteAction(id));
+    }
   };
   return (
     <div>
